Restore eight regressed ERP requirements across reception, adjustments, and quotes
- Approval letter skip UI and defaults endpoint; fix cost confirmation loops
- Fractional quantities end-to-end with decimal migration and validation
- Adjustments grand total, form field policies wiring, notification load beep
- Quote print uses free description (written_items); fix notification login tests

+++ app/Http/Controllers/Adjustments/AdjustmentsController.php
@@ -113,7 +113,7 @@ class AdjustmentsController extends Controller
 
     public function updateItemQty(UpdateAdjustmentItemQtyRequest $request, CaseRecord $case, BomItem $bomItem): JsonResponse
     {
-        $bom = $this->adjustmentsService->updateItemQty($case, $bomItem, (int) $request->validated('qty'));
+        $bom = $this->adjustmentsService->updateItemQty($case, $bomItem, (float) $request->validated('qty'));
 
         return response()->json([
             'message' => 'تم تحديث كمية البند.',
